docs: Remove use of 'jqXHR' and 'mediaWiki' jsduck aliases
The `@alternateClassName` feature does not map cleanly to
JSDoc, so reduce its use to make the migration smoother.

Also, it was my intention to only use `@alternateClassName`
as a way to overcome limitations in JSDuck and/or as a stop-gap
to document hacky or overly clever code, not as a general way
to just randomly have two names for the same thing within docs.

* jqXHR: Just use jQuery.jqXHR instead. Also add missing
  inheritence to Promise.

* mediaWiki: Just use mw instead. It's already discouraged
  in real code, we really don't need to allow it within
  doc typehints.

Bug: T138401
Change-Id: I915f747e440bad8f9bb2c11f35abc32ebdb69247