Rename ipb_address to ipb_address_unique
This is how it's named in WMF production and the Postgres. Converging
names would help us to keep everything consistent, for example if we
need to force this index in the code.

For changing this index in Sqlite, we don't need to make a temp table as
this only touches the index and not the table structure.

Bug: T250071
Change-Id: I0838753b809680af3ee06d9c4f6d21d95b8982cc