[WIP] Workaround PHP 7.4>= memory leak in tests
On PHP 7.4 and above, a memory leak can be observed in
ServiceContainer::loadWiringFiles when running PHPUnit tests. This patch aims to
work around the leak by changing
MediaWikiIntegrationTestCase::installMockMwServices to cache and reuse the
original service wirings if no config overrides were supplied. This
significantly reduces the amount of times that the wiring closures need to be
included; on PHP 8, it caused memory usage to drop to 894.25 MB from 4+ GB when
running the PHPUnit tests for MediaWiki core.

However, it seems to be causing some test failures and so it is currently a WIP.

Bug: T247990
Change-Id: I24971221b8accf78d61479d286e907c1111212a0