xml: Move condition check of var before creating XmlSelect object
The `$selected` var could be null and in this case, return the default,
an empty string. The null check was after creating the XmlSelect
object which seemed out of place as after setting the variable to
the default, it's not used anywhere below.

Actually, the check should happen above and set it to it's default if
need be before creating the object below.

Change-Id: I6b69a7780c9f4a9ae283db8fa90b9cee76954d00