Move french space armoring after doBlockLevels
This was moved before doBlockLevels in c23cb2c when it used to be a more
general cleanup that conflicted with it.  Since 02ff859 that no longer
seems necessary.

Further, this gives the consistency of armoring inside wikilinks and
avoids some unnecessary row definition list colons being armored.

Maybe this should be moved to the end of internalParseHalfParsed after
language conversion but before tidying, where the other call to the
sanitizer happens, Sanitizer::normalizeCharReferences.  But let's change
things one step at a time and start by moving it here before we try to
move it further.

The goal here being to make french space armoring a post-processing
step, rather than an intermediate transformation, more consistent with
how Parsoid will/wants to do it.

Bug: T197879
Change-Id: I3eae3719ab8fb50b7996d4fd8a9fa0d5ca250023