Encase 'actor_id' value in aggregate function
In PostgreSQL, all selected columns must be present in the GROUP BY clause except if their values are derived from computation (of aggregation function). In this query, 'user_id' and 'user_name' are already in the clause while 'user_real_name' and 'timestamp' make use of aggregate functions, therefore not required to be in the clause. 'actor_id' does not use any aggregate function and is missing in the clause which triggers the failures.

I did not test this with real PostGres database, but it works correctly with MariaDB and I believe it'll fix this issue too.

Bug: T251599
Change-Id: I10c8d84dbb58ea5258671126a6b05ebf93bab312