mediawiki.base: Let exists() work even in lang=qqx
== Background

In all languages, when a message has not been bundled by the server (because it
was not declared in the module, or because it doesn't exist), then
mw.message().exists() returns false, and if you fetch the message anyway,
you get "⧼thekey⧽" as fallback rendering.

This changed as part of bb084a3573 (I2266ec0), which had as intent to
provide the "display message parameter" feature for lang=qqx, the same
as the PHP rendering did already.

However, it had as side-effect that it discarded the actual value
bundled by the server, and it also discarded the fact of whether
it was bundled at all.

This changed again as part of 616b6176db (I3f2a6ce), which restored
the part where the original value is honoured (if different).

This commit also restores the part about honouring message existence.

== Why

As part of T250045, we want to use lang=qqx by default in QUnit
context. In trying that out, I noticed some of Minerva's tests
started to fail.

> Minerva pageIssues: insertBannersOrNotice() should add "learn more"
> Expected: ⧼skin-minerva-issue-learn-more⧽
> Actual: (skin-minerva-issue-learn-more)

The test asserts whether the message key is renderered, specifically
it does that by not loading the message, and then expecting it
to render as "⧼thekey⧽". If qqx ignores message existence, then
this would not be forward and backward compatible and likely require
many changes.

Also more generally, it seems useful to preserve the ability to
distinguish whether a message was registered and transferred
correctly to the client - whilst also having the benefit of a
rendering that is both language-neutral and cannot change over time
due to changes in translations.

Bug: T250045
Change-Id: Ie9eed5c9e886901d6867eff5439848436ad822a8