resourceloader: optimise 'nextCssBuffer' to not need an 'active' property
There can be more than one buffer in-flight. The flush function
has the buffer object that it will flush bound to it. That way,
there is no reliance on the shared 'nextCssBuffer' variable, which
can change in the meantime.

To avoid closing the wrong one, the callback used a property to track
the 'active' state. If another buffer was started and assigned to
'nextCssBuffer' in the meantime, then the change to 'active' would
never be seen. This is good, but also a wasted operation and property.

Rather than that, simply null out the nextCssBuffer variable
directly.

In addition to needing fewer properties and fewer operations,
this also means that we no longer hold on indefinitely in JS
memory to the last CSS buffer. Instead, it will always end up
cleanly in a 'null' state when all is done. Previously it would
end in the state of 'nextCssBuffer' holding the last buffer
with 'active=false'.

Also:

* The active property had three states: null, false, true.
  The latter two are explained above, but null was special.
  The null value was basically just a a way to track within
  the addEmbeddedCSS() function whether it created a new one
  just now, to know whether to call rAF() to schedule the flush.

  There was no need to track that state, as rAF() can simply
  be called directly from the same conditional block.

Change-Id: Ic78529b33afdd9ccf25b4abc846a000e7399fcc1