resourceloader: remove obsolete media=all optimisation in addLink()
This was an optimisation to reduce DOM property assignments, which
come with a slightly higher cost than normal property assignments.

However, loading raw urls with <link rel=stylesheet> is only done
in debug mode so this optimisation isn't very useful here.

Also, in general the server responds with an array of urls
and the responses are wrapped with `@media` as needed. The
model of passing 'media' as a key to the client hasn't been
used since at least 2011, possibly earlier.

Commits from around that time: ff594d3b0a, and 15e5193a19.

Change-Id: Ied08b3522e9db027d479e775b1b4b9348e9e76cc