registration: Restrict properties in "config" in manifest_version 2
This allows for the schema to spot typos because there's now a whitelist of
properties. The main thing that might get caught by this is documentation
using "@something" keys, but all of that should be under "description"
now.

I sent patches to all extensions/skins in Gerrit so they'll pass validation
with the new schema: <https://gerrit.wikimedia.org/r/q/topic:%22fix-config-validation%22>.

Bug: T252589
Change-Id: I94d96b8a66b9209fa32726c2a2166308536f6c8a