AbstractBlock: Fix documentation for parseTarget
parseTarget is documented as accepting an int as the $target param,
but it will always treat it as a string. Any int would be interpreted
as the user name (string)$target, and any caller passing user names
would be passing a string, since most user names can't be expressed
as an int.

This fixes the documentation to allow only string or User.

Change-Id: I3dd20b91bbe8da7fee67bd336c7e973a22cf082d