editpage: Fix misleading "OOUI\FieldLayout" type hint
It looks like this was meant to document what the method actually
returns. But this is misleading. Subclasses (which exist) are not
required to return this.

Later the array is fed to \OOUI\HorizontalLayout which is documented
to accept "Widget[]|Layout[] … Widgets or other layouts to add to
the layout." This is what I copied.

Both share a common base class, \OOUI\Element, which I use as the
actual type hint.

Change-Id: I812cb388bef728b12ec2bfdcd2164e1c2cec13ae