media: Fix mismatching/incomplete PHPDocs related to metadata
* Use false instead of bool in PHPDocs, because that's the only bool
value that's allowed.

* This patch also fixes DjVuHandler::getPageText() not returning
strings, but XML objects. This kind of "worked" because all consuming
code magically casts these to strings. But this is an actual violation
of the contract of the method. This is also why the test was doing this
weird (string) cast, instead of actually testing the type of the return
value.

Change-Id: I00db6b910f1de6d37a80543b8a5dd5ea3bab3c76