SpecialBlock: Fix flag set on block if $wgBlockAllowsUTEdit is false
The flag `ipb_allow_usertalk` should be false for every block if
$wgBlockAllowsUTEdit is false. However, since 02cb7aefef02d, the flag
is true if the config is false - if the block is made via
SpecialBlock::processForm.

This is particularly needed ahead of T189073, in which we refactor
SpecialBlock::processForm into a blocking service. As we update
extensions to use this service, we don't want to change the
`ipb_allow_usertalk` flags on their blocks on wikis where
$wgBlockAllowsUTEdit is false.

Bug: T252892
Change-Id: I6f38c5c62bcb2216ff5bf0a7942d678bdd1a05f8