Treat languages and toolbox like other sidebar links
This will require some changes to individual skins but the toolbox
and languages portals should be included in the SidebarBeforeOutput
hook. Otherwise two new hooks for adding languages and toolboxes
will be needed.

The mapInterwikiToLanguage,
buildNavUrls, buildFeedUrls and getLanguages functions are moved
from SkinTemplate up to Skin. The bodies of these functions remain the
same with the following changes:

* slight readability tweak inside getLanguages and buildFeedUrls
to simplify the return value.
* buildNavUrls is modified to use a local
variable for thispage rather than the public variable on SkinTemplate.
* title and user cached inside buildNavUrls rather than calling getTitle
and getUser multiple times
* Use the hook service rather than Hooks::run

This will allow us to DRY up every skin which sets these sidebar portals
on when they are not set as well as allow extensions to add items to
the toolbox and language portals just like any other sidebar portal.

https://codesearch.wmflabs.org/skins/?q=TOOLBOX&i=nope&files=&repos=

Note: We will need to make modifications to all skins to remove
their existing hardcoding of these portals.

Bug: T252917
Change-Id: Ie33f85a77299da8f3b98ab8b363761d1f8277708