Add caption to always suppressing
In brief, the BlockLevelPass looks at opening and closing tags on a line
to determine whether it should do paragraph wrapping.  The blockElems
want to stop wrapping when opened and start again when closed.  The
antiBlockElems want the opposite, to start when they're opened and stop
when closed.  "table" is a blockElems and "td"|"th" are anitBlockElems
so that content found in the interstitial spaces of tables are never
paragraph wrapped.

That means that, to date, "caption" elements are always found in a place
where paragraph wrapping is always suppressed and so adding them to that
set won't change any test results.  However, a new test is added to spec
out this behaviour.

In the legacy parser, "captions" are always found in the right place
because handleTables runs at an earlier stage.  In Parsoid, however, the
treebuilder is relied on to close table cells [0] so when we get to the
token stream paragraph wrappping pass, "caption"s are found in table
cells and therefore get wrapped, even though the treebuilder is about to
be induced to close the cell before opening the caption.

Therefore, in Parsoid, the fix would require us to make captions always-
suppressing to match the legacy parser behaviour.  Thus, this change
here is just to keep these lists [1] consistent between the two
parsers.

[0] https://github.com/wikimedia/parsoid/blob/5e11a3f3908932e097c42d3e205c78721a84dedf/src/Wt2Html/TokenizerUtils.php#L138-L151
[1] https://github.com/wikimedia/parsoid/blob/5e11a3f3908932e097c42d3e205c78721a84dedf/src/Wt2Html/TT/ParagraphWrapper.php#L71-L78

Bug: T210647
Change-Id: I8ccefd69d47dca740f50924b235dffa3873d1f99