[EditPage] Move AS_* status constants to a new interface
Interface is @internal and is implemented by EditPage; will be used in
the future when the backend is split and common status values are needed
for the different classes.

Doing separately to avoid cluttering the main work - just moves the
constant definitions. Since EditPage implements the new interface, any
existing EditPage::AS_* references will continue to work.

Bug: T20654
Change-Id: I71dafec2579b4a42dd08f3fc7e569e28c14d8753