resourceloader: Add internal handling for debug=2
The canonical way to enable debug mode is and will remain
through via 'debug=true'.

During the transition debug=2 will opt you in to the experimental
newer way. Anything that needs to be pinned to the old way for
compat can already start doing so by using debug=1 explicitly.

Once v2 is "ready", the default will flip and debug=1 will remain
for the foreseeable future to trigger the legacy behaviours.

Bug: T85805
Change-Id: Ieaf04e0c289646dd5d5b027b4f1f8278167b2d57