jquery.tablesorter: Return to initial sort order on third click
Add a third sort order, which returns the column to the initial
order. Use the title and the sort icon to indicate that a third
click will return to the original order.

If sorting multiple columns, the clicked-on column will use the
initial sort order, but restricted by the current orders of any
other sorted columns.

If cells were initially merged, but then became unmerged due to
sorting, this patch does not re-merge them. That could be added
in the future.

Bug: T226697
Change-Id: I11cd7c734e2ff9679db0cdd431a33d040daf2e32