MediaWikiServices: All logic for services should be in ServiceWiring
Move handling of the different services that are created with
NameTableStoreFactory to ServiceWiring, so that MediaWikiServices
can just call $this->getService() and nothing else.

For objects created from specs via ObjectFactory, services are retrieved
via ContainerInterface::get( 'ServiceName' ), an alias for ::getService.
Since the different stores created with NameTableStoreFactory were not
individually defined as services, they would not have been usable
with an ObjectFactory spec.

Change-Id: I8f862efeffe91a8ca1c4cbc48a5c429d6c78973c