Links added via SidebarBeforeOutput are exposed to skins using BaseTemplate
* Follow up for Ie33f85a. SidebarBeforeOutput hook was allowed to modify
Toolbox and Languages links but they were not actually appended if any
of 'TOOLBOX' or 'LANGUAGES' magic string is used, because these strings
are already processsed earlier.
* This patch ensures these hooks modifications are merged to the final
content of the section specified with the magic string.
* Also improve documentation of 'SEARCH' magic string. Unlike other two
strings, this string is not recognized out-of-the-box. It works only
when a skin explicity added support for it, otherwise content
added to it will not be displayed.
* Several new line formatting changes for readability.

Bug: T252917
Change-Id: Ie10a59048cdf882bad811d3c8fae386d90e93867