Stop throwing an exception in UploadStash::getExtensionForPath
The exception serves no purpose, and can only really be triggered via
a test. The API prevents no file extension at all, as does UW js.

This function (for whatever reason, probably a seperate bug) cannot
get the extension from a stashed stl file (seems to work fine for
other types).

With what/how it's actually used, it doens't really matter if
we can't get the extension, we get it by more robust methods later
on.

This partially reverts 0a82600a27. Before the changes in that commit,
the exception was unreachable.

Bug: T254078
Change-Id: I0a7bd13fe8e08c7d4a75b4a3709661dbbf53d6cb