Don't suppress debug log lines for queries when the debug toolbar is enabled
Log lines for SQL queries were not shown in either the debug toolbar's
debug log, or in the file configured by $wgDebugLogFile. I think it is
useful to have SQL queries both in a timeline context in the debug log,
and separately in table format.

This is a very similar complaint to I8edb41aa1d8236047b80830, which
affected deprecation messages.

Change-Id: Ie0ca7bf8394d3008e3f9e950c201da1a96276fea