Revert "title: Mark Title::getPrefixedDBKey and ::getPrefixedText as tainted"
This reverts commit a9ba5f98a4d6548d7ca5884d3f5e55cc955d1a8b.

Reason for revert: these props are unsafe iff the instance they belong to is unsafe. Marking them as *always* unsafe turned out to be a source of false positives which doesn't compensate the gain. Note: merging this patch will likely make phan check in a few repos with "UnusedSuppression ...". It can be fixed by simply removing the suppression it complains about.

Change-Id: I6e2c13400f51b42269d7d07976809a986eb67b05