mediawiki.misc-authed-pref: Limit dblClickEdit to 'view' action
Follows-up c656980498, which led to this also being loaded when
the rightClickEdit feature is loaded (which is loaded on all actions).
It's a tiny opt-in script so no issue there, but we do need to make
sure it's self-contained in terms of what it is meant to do, and not
rely purely on when it is included in a bundle send to the client.

Change-Id: Ie13a4de90c0ef3cc6673bd352bc25d1e02df8dc4