CloneDatabase, add domain alias when prefix is changed
I was having a problem writing an integration test for
I3801b4d1f4a876a820c28f4541ec0802d886b7eb as part of T254283.

After much debugging I realized that 2 different db domains were being
used by the tests, resulting in inconsistent queries.
This is partly due to some recent changes in Wikbase code that no longer
always passes false around for the local domain, which was T252019
Ie8822ebd3661a21cbe475f75cae2615cc2b206c6
I35fa856b0b0591e814ddbfda2fed04322c6dd63f

When unit tests use a test prefix, the code in resolveDomainInstance no
longer resolves the db name string to the correct domain (with the
prefix) instead using the orig domain (with no test prefix).

This is due to the $domain === $this->localDomain->getId() check as for
a db name of "default" the equiv string id is now "default-unittest_"

This patch adds aliases to LB and LBF for the domain when a prefix is
used.

Aaron should probably confirm if this is the best place for this code to
live? Or if it should actually live in the test base class?

Bug: T254283
Bug: T252019
Change-Id: I4615c32cefc2e1af685137cb96c37e3813bdce75